requirements.txt
django
djangorestframework
psycopg2-binary
django-environ
djangorestframework-gis
django-cors-headers
django-filter